add Flyway

+++ src/main/java/com/fortech/academy/library/controller/CreateRoomRequest.java
@@ -1,10 +1,7 @@
 package com.fortech.academy.library.controller;
 
-import jakarta.persistence.Column;
-
 public class CreateRoomRequest {
-
-    private Long hotelId;
+    private int hotelId;
     private int roomNumber;
 
     private String roomType;
